WIP: skip docker tests if docker is not found
Makes easier to run molecule test suite on platforms where docker
is not available.
