+++ molecule/test/conftest.py
@@ -19,6 +19,7 @@
 #  DEALINGS IN THE SOFTWARE.
 
 import contextlib
+import distutils
 import os
 import random
 import string
@@ -28,9 +29,19 @@ import pytest
 from molecule import config
 from molecule import logger
 from molecule import util
+from molecule.driver.docker import Docker
 from molecule.scenario import ephemeral_directory
 
 LOG = logger.get_logger(__name__)
+
+
+@util.lru_cache()
+def has_docker():
+    try:
+        Docker().sanity_checks()
+    except:
+        return False
+    return True
 
 
 @pytest.helpers.register
@@ -151,3 +162,140 @@ def reset_pytest_vars(monkeypatch):
     for var_name in tuple(os.environ):
         if var_name.startswith('PYTEST_'):
             monkeypatch.delenv(var_name, raising=False)
+
+
+@pytest.fixture
+def patched_docker_sanity_check(mocker):
+    return mocker.patch(
+        'molecule.driver.docker.Docker.sanity_checks', return_value=True
+    )
+
+
+@pytest.fixture
+def skip_test(request, driver_name):
+    msg_tmpl = (
+        "Ignoring '{}' tests for now"
+        if driver_name == 'delegated'
+        else "Skipped '{}' not supported"
+    )
+    support_checks_map = {
+        'digitalocean': supports_digitalocean,
+        'docker': supports_docker,
+        'ec2': supports_ec2,
+        'gce': supports_gce,
+        'hetznercloud': lambda: at_least_ansible_28() and supports_hetznercloud(),
+        'linode': lambda: at_least_ansible_28() and supports_linode(),
+        'openstack': supports_openstack,
+        'vagrant': supports_vagrant_virtualbox,
+        'delegated': demands_delegated,
+    }
+    try:
+        check_func = support_checks_map[driver_name]
+        if not check_func():
+            pytest.skip(msg_tmpl.format(driver_name))
+    except KeyError:
+        pass
+
+
+def get_docker_executable():
+    return distutils.spawn.find_executable('docker')
+
+
+def get_vagrant_executable():
+    return distutils.spawn.find_executable('vagrant')
+
+
+def get_virtualbox_executable():
+    return distutils.spawn.find_executable('VBoxManage')
+
+
+@pytest.helpers.register
+def supports_docker():
+    return get_docker_executable()
+
+
+@pytest.helpers.register
+def supports_digitalocean():
+    try:
+        # ansible >=2.8
+        # The _digital_ocean module is deprecated, and will be removed in
+        # ansible 2.12. This is a temporary fix, and should be addressed
+        # based on decisions made in the related github issue:
+        # https://github.com/ansible/molecule/issues/2054
+        from ansible.modules.cloud.digital_ocean._digital_ocean import HAS_DOPY
+    except ImportError:
+        # ansible <2.8
+        from ansible.modules.cloud.digital_ocean.digital_ocean import HAS_DOPY
+
+    env_vars = ('DO_API_KEY',)
+
+    return _env_vars_exposed(env_vars) and HAS_DOPY
+
+
+@pytest.helpers.register
+def supports_ec2():
+    from ansible.module_utils.ec2 import HAS_BOTO3
+
+    env_vars = ('AWS_ACCESS_KEY', 'AWS_SECRET_ACCESS_KEY')
+
+    return _env_vars_exposed(env_vars) and HAS_BOTO3
+
+
+@pytest.helpers.register
+def supports_gce():
+    from ansible.module_utils.gcp import HAS_GOOGLE_AUTH
+
+    env_vars = ('GCE_SERVICE_ACCOUNT_EMAIL', 'GCE_CREDENTIALS_FILE', 'GCE_PROJECT_ID')
+
+    return _env_vars_exposed(env_vars) and HAS_GOOGLE_AUTH
+
+
+@pytest.helpers.register
+def supports_hetznercloud():
+    pytest.importorskip('hcloud')
+
+    env_vars = ('HCLOUD_TOKEN',)
+
+    return _env_vars_exposed(env_vars)
+
+
+@pytest.helpers.register
+def supports_linode():
+    from ansible.modules.cloud.linode.linode_v4 import HAS_LINODE_DEPENDENCY
+
+    env_vars = ('LINODE_ACCESS_TOKEN',)
+
+    return _env_vars_exposed(env_vars) and HAS_LINODE_DEPENDENCY
+
+
+@pytest.helpers.register
+def supports_vagrant_virtualbox():
+    return get_vagrant_executable() or get_virtualbox_executable()
+
+
+@pytest.helpers.register
+def supports_openstack():
+    pytest.importorskip('shade')  # Ansible provides no import
+
+    env_vars = (
+        'OS_AUTH_URL',
+        'OS_PASSWORD',
+        'OS_REGION_NAME',
+        'OS_USERNAME',
+        'OS_TENANT_NAME',
+    )
+
+    return _env_vars_exposed(env_vars)
+
+
+@pytest.helpers.register
+def demands_delegated():
+    return pytest.config.getoption('--delegated')
+
+
+def _env_vars_exposed(env_vars, env=os.environ):
+    """Check if environment variables are exposed and populated."""
+    for env_var in env_vars:
+        if env_var not in os.environ:
+            return False
+        return os.environ[env_var] != ''
