+++ molecule/test/functional/conftest.py
@@ -18,8 +18,6 @@
 #  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
 #  FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 #  DEALINGS IN THE SOFTWARE.
-
-import distutils.spawn
 import os
 import pkg_resources
 import shutil
@@ -31,19 +29,11 @@ import sh
 from molecule import logger
 from molecule import util
 
-from ..conftest import change_dir_to
+from ..conftest import change_dir_to, skip_test
 
 LOG = logger.get_logger(__name__)
 
 IS_TRAVIS = os.getenv('TRAVIS') and os.getenv('CI')
-
-
-def _env_vars_exposed(env_vars, env=os.environ):
-    """Check if environment variables are exposed and populated."""
-    for env_var in env_vars:
-        if env_var not in os.environ:
-            return False
-        return os.environ[env_var] != ''
 
 
 @pytest.fixture
@@ -63,32 +53,6 @@ def with_scenario(request, scenario_to_test, driver_name, scenario_name, skip_te
             options = {'driver_name': driver_name, 'all': True}
             cmd = sh.molecule.bake('destroy', **options)
             pytest.helpers.run_command(cmd)
-
-
-@pytest.fixture
-def skip_test(request, driver_name):
-    msg_tmpl = (
-        "Ignoring '{}' tests for now"
-        if driver_name == 'delegated'
-        else "Skipped '{}' not supported"
-    )
-    support_checks_map = {
-        'digitalocean': supports_digitalocean,
-        'docker': supports_docker,
-        'ec2': supports_ec2,
-        'gce': supports_gce,
-        'hetznercloud': lambda: at_least_ansible_28() and supports_hetznercloud(),
-        'linode': lambda: at_least_ansible_28() and supports_linode(),
-        'openstack': supports_openstack,
-        'vagrant': supports_vagrant_virtualbox,
-        'delegated': demands_delegated,
-    }
-    try:
-        check_func = support_checks_map[driver_name]
-        if not check_func():
-            pytest.skip(msg_tmpl.format(driver_name))
-    except KeyError:
-        pass
 
 
 @pytest.helpers.register
@@ -245,23 +209,6 @@ def verify(scenario_name='default'):
     pytest.helpers.run_command(cmd)
 
 
-def get_docker_executable():
-    return distutils.spawn.find_executable('docker')
-
-
-def get_vagrant_executable():
-    return distutils.spawn.find_executable('vagrant')
-
-
-def get_virtualbox_executable():
-    return distutils.spawn.find_executable('VBoxManage')
-
-
-@pytest.helpers.register
-def supports_docker():
-    return get_docker_executable()
-
-
 def at_least_ansible_28():
     """Ensure current Ansible is >= 2.8."""
     try:
@@ -273,82 +220,3 @@ def at_least_ansible_28():
     except ImportError as exception:
         LOG.error('Unable to parse Ansible version', exc_info=exception)
         return False
-
-
-@pytest.helpers.register
-def supports_linode():
-    from ansible.modules.cloud.linode.linode_v4 import HAS_LINODE_DEPENDENCY
-
-    env_vars = ('LINODE_ACCESS_TOKEN',)
-
-    return _env_vars_exposed(env_vars) and HAS_LINODE_DEPENDENCY
-
-
-@pytest.helpers.register
-def supports_vagrant_virtualbox():
-    return get_vagrant_executable() or get_virtualbox_executable()
-
-
-@pytest.helpers.register
-def demands_delegated():
-    return pytest.config.getoption('--delegated')
-
-
-@pytest.helpers.register
-def supports_digitalocean():
-    try:
-        # ansible >=2.8
-        # The _digital_ocean module is deprecated, and will be removed in
-        # ansible 2.12. This is a temporary fix, and should be addressed
-        # based on decisions made in the related github issue:
-        # https://github.com/ansible/molecule/issues/2054
-        from ansible.modules.cloud.digital_ocean._digital_ocean import HAS_DOPY
-    except ImportError:
-        # ansible <2.8
-        from ansible.modules.cloud.digital_ocean.digital_ocean import HAS_DOPY
-
-    env_vars = ('DO_API_KEY',)
-
-    return _env_vars_exposed(env_vars) and HAS_DOPY
-
-
-@pytest.helpers.register
-def supports_ec2():
-    from ansible.module_utils.ec2 import HAS_BOTO3
-
-    env_vars = ('AWS_ACCESS_KEY', 'AWS_SECRET_ACCESS_KEY')
-
-    return _env_vars_exposed(env_vars) and HAS_BOTO3
-
-
-@pytest.helpers.register
-def supports_gce():
-    from ansible.module_utils.gcp import HAS_GOOGLE_AUTH
-
-    env_vars = ('GCE_SERVICE_ACCOUNT_EMAIL', 'GCE_CREDENTIALS_FILE', 'GCE_PROJECT_ID')
-
-    return _env_vars_exposed(env_vars) and HAS_GOOGLE_AUTH
-
-
-@pytest.helpers.register
-def supports_hetznercloud():
-    pytest.importorskip('hcloud')
-
-    env_vars = ('HCLOUD_TOKEN',)
-
-    return _env_vars_exposed(env_vars)
-
-
-@pytest.helpers.register
-def supports_openstack():
-    pytest.importorskip('shade')  # Ansible provides no import
-
-    env_vars = (
-        'OS_AUTH_URL',
-        'OS_PASSWORD',
-        'OS_REGION_NAME',
-        'OS_USERNAME',
-        'OS_TENANT_NAME',
-    )
-
-    return _env_vars_exposed(env_vars)
